pipeline_server/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
openpyxl
odfpy
chardet
rapidfuzz
phonenumbers
python-dateutil
arabic-reshaper
python-bidi
gender-guesser
scikit-learn
loguru
sqlalchemy